update : modularized firebaseApi
