+++ components/auth/AuthRegisterForm.tsx
@@ -11,7 +11,7 @@ import { useAuthCtx } from "../../utils/contexts/auth/AuthHook";
 import { MainNetworkResponse } from "../../utils/data/Main";
 import { UserModel } from "../../utils/data/models/UserModel";
 import { waitFor } from "../../utils/helpers/DelayHelpers";
-import { fbUserRegister } from "../../utils/services/network/FirebaseApi";
+import { fbUserRegister } from "../../utils/services/network/FirebaseApi/UserModules";
 import MainTextInput from "../input/MainTextInput";
 import { StatusPlaceholderAction } from "../placeholder/StatusPlaceholder";
 import { AuthFormProps } from "./AuthPanel";
@@ -160,7 +160,7 @@ function AuthRegisterForm({
           // if error
           if (resp.status === "error")
             return actionError(resp as MainNetworkResponse<FirebaseError>, () =>
-              onSubmit(data),
+              onSubmit(data), 
             );
           // if success
           if (resp.status === "success")
