+++ components/profile/ProfileSetupForm.tsx
@@ -17,7 +17,7 @@ import { waitFor } from "../../utils/helpers/DelayHelpers";
 import { transitionPullV } from "../../utils/helpers/UiTransitionHelpers";
 import { useNetworkAction } from "../../utils/hooks/NetworkActionHook";
 import { scrollToTop } from "../../utils/hooks/RouteChangeHook";
-import { fbUserUpdate } from "../../utils/services/network/FirebaseApi";
+import { fbUserUpdate } from "../../utils/services/network/FirebaseApi/UserModules";
 import MainTextAreaInput from "../input/MainTextAreaInput";
 import MainTextInput from "../input/MainTextInput";
 import GradientBackground from "../main/GradientBackground";
